IBM1130: Fix GUI initialization
Long running Heisenbug in IBM1130 during RegisterSanityCheck tests,
where the simulator appears to hang indefinitely and eventually times
out. The GUI thread (Pump) is prematurely killed, leaving GUI resources
in a partially initialized state, that subsequently puts the Windows C
runtime into a wacky state and prevents the executable from exiting.

Note: Creating a thread != thread immediately executes, as was the tacit
assumption in the code. Threads, like processes, are scheduled and do
not execute determinisitically.

Fix: Use a critical section and condition variable to signal the main
thread that all GUI resources have been acquired and that Pump is about
to enter the message loop.

Other changes:

ibm1130:
  - Fix Windows type warnings for better Win64 compatibility.
  - Reduce compiler warnings.
  - Windows: Add ibm1130.rc resource file as a simulator source in the
    CMake build.

sim_defs.h:
  - Add a SIM_UNUSED_PARAM macro to tamp down warnings about unused
    arguments. See comment for usage.

Notes for the IBM1130 maintainer(s):
- ibm1130_cr.c
  - feedcycle(): Suggest changing nxread, nread type to size_t
  - pcr_thread(): nrcvd, nwritten types -> DWORD (upgrade from long)
  - pcr_cmd(): nrcvd, nwritten types -> DWORD (upgrade from long)

+++ Ibm1130/ibm1130_fmt.c
@@ -290,7 +290,7 @@ int     bContinue;                                          /* true if continue 
 
 const char* EditToWhitespace(char *p_pszEdit, int width)
 {
-int     iI;                                                 /* work integer */
+size_t  iI;                                                 /* work integer */
 int     iPos;                                               /* work integer for settings tab stops */
 char    pszLine[MAXLINE];                                   /* source line */
 char    pszWork[WORKSZ];                                    /* work buffer */
@@ -307,7 +307,7 @@ char    pszWork[WORKSZ];                                    /* work buffer */
         iI = 0;                                             /* output index */
         iPos = width + 1;                                   /* first tab position */
         while (iPos < 80) {                                 /* fill array up to but not including position 80 */
-            gaiPlainTabs[iI++] = iPos;
+            gaiPlainTabs[iI++] = (short) iPos;
             iPos += width;
         }
         gaiPlainTabs[iI] = 0;                               /* mark end of array */
